move some code to separated methods and modules
- move get_cluster_proxies and update_container_with_proxy_env functions
  to new module ocs_ci.helpers.proxy
- move code for configuring proxy to new method
  configure_allowed_domains_in_proxy
